refactor(effects): 修改模糊效果的尺寸处理逻辑
- 将EBBlur类中的resize方法参数从int改为float类型
- 在MyDraw类中同步修改blurRect方法的调用参数传递
- 统一DrawEffect接口的resize方法签名以支持浮点数尺寸
- 调整窗口绘制层级避免模糊效果遮挡问题

+++ src/modtools/ui/effect/EBBlur.java
@@ -168,15 +168,14 @@ public class EBBlur implements DrawEffect {
 		return new Shader(vertexShader, fragmentShader);
 		// return null;
 	}
-
-	public void resize(int width, int height) {
+	public void resize(float width, float height) {
 		blurScl = E_Blur.scale_level.getInt();
 		if (blurScl == 0) blurScl = 1;
 		width /= blurScl;
 		height /= blurScl;
 
-		buffer.resize(width, height);
-		pingpong.resize(width, height);
+		buffer.resize((int) width, (int) height);
+		pingpong.resize((int) width, (int) height);
 
 		blurShader.bind();
 		blurShader.setUniformf("size", width, height);
